Replace updating the original response with the proper interaction response
Before realizing there was a response type for that, the bot was sending an acknowledge and a response update for the original response separately. This complicated some of the responses and also limited the maximum possible expiration time to 15 minutes (at which point, the original message token would expire, and the message could no longer be edited).